Migration!
Started work on migration from 23' to 24'
Figured out the REV migration thanks to their awesome changelog, but having some trouble with the CTRE migration for pigeon2.
Looking into it currently, but might be having a issue we've had before were we defined the wrong pigeon2.
Also, build.gradle error :sob: